[Var] foobar2000: Update SDK to SDK-2015-08-03.zip.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@5468 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ include/foobar2000sdk/pfc/base64.cpp
@@ -37,7 +37,7 @@ namespace pfc {
 		static const char alphabet[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
 		t_uint8 alphabetRev[256];
 		for(t_size walk = 0; walk < PFC_TABSIZE(alphabetRev); ++walk) alphabetRev[walk] = 0xFF;
-		for(t_size walk = 0; walk < PFC_TABSIZE(alphabet); ++walk) alphabetRev[alphabet[walk]] = (t_uint8)walk;
+		for(t_size walk = 0; walk < PFC_TABSIZE(alphabet); ++walk) alphabetRev[ (unsigned) alphabet[walk]] = (t_uint8)walk;
 		const t_size textLen = strlen(text);
 
 		if (textLen % 4 != 0) throw pfc::exception_invalid_params();
